movie_project/requirements.txt
Click==7.0
Flask==1.0.2
Flask-Redis==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysql-connector-python==8.0.13
protobuf==3.6.1
PyMySQL==0.9.3
redis==3.1.0
six==1.12.0
SQLAlchemy==1.2.16
Werkzeug==0.14.1
WTForms==2.2.1
